Create dedicated module.
